+++ examples/data-sources/meraki_organization_branding_policies_priorities/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_organization_branding_policies_priorities" "example" {
+  organization_id = "123456"
+}