+++ examples/resources/meraki_organization_branding_policies_priorities/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_organization_branding_policies_priorities.example "<organization_id>"