+++ examples/resources/meraki_organization_branding_policies_priorities/resource.tf
@@ -1,0 +1,4 @@
+resource "meraki_organization_branding_policies_priorities" "example" {
+  organization_id     = "123456"
+  branding_policy_ids = ["123"]
+}