+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add `meraki_wireless_ssids` data source
 - Add `meraki_organization_early_access_features` data source
 - Add `meraki_organization_branding_policy` resource and data sources
+- Add `meraki_organization_branding_policies_priorities` resource and data source
 
 ## 0.1.2